% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked_points.R
\name{get_in_list}
\alias{get_in_list}
\title{get in_list for get_linked_points_scalable}
\usage{
get_in_list(lp_points, net)
}
\arguments{
\item{lp_points}{as returned in the "lp" named output of get_lp_points}

\item{net}{sf data.frame NHDPlus Flowlines}
}
\description{
Allows atomic access to in_list creation logic for calling functions used
in get_linked_points in a workflow.
}
