% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_levelpaths.R
\name{get_length_per_hu}
\alias{get_length_per_hu}
\title{get_length_per_hu}
\usage{
get_length_per_hu(net, wbd, simp)
}
\arguments{
\item{net}{nhdplus}

\item{wbd}{wbd}

\item{simp}{simplification}
}
\description{
get_length_per_hu
}
