% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked_points.R
\name{get_lp_points}
\alias{get_lp_points}
\title{Get LevelPath Points}
\usage{
get_lp_points(hu_lp, net, wbd, exclude)
}
\arguments{
\item{hu_lp}{data.frame as returned by `match_levelpaths`}

\item{net}{sf data.frame ...}

\item{wbd}{sf data.frame ...}

\item{exclude}{character vector of HUs to exclude from consideration}
}
\description{
finds
}
