% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked_points.R
\name{get_na_outlets_coords}
\alias{get_na_outlets_coords}
\title{Get NA outlet coords}
\usage{
get_na_outlets_coords(na_points, net)
}
\arguments{
\item{na_points}{data.frame as returned in the "na" named return from 
get_lp_points.}

\item{net}{sf data.frame NHDPlus Flowlines}
}
\value{
sf data.frame with outlet locations
}
\description{
For HUs that have a levelpath but no intersection 
returns the outlet of the interior flowline.
}
