% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_levelpaths.R
\name{par_match_levelpaths}
\alias{par_match_levelpaths}
\title{Parallel match levelpaths}
\usage{
par_match_levelpaths(net, wbd, simp, cores, temp_dir = "temp/",
  out_file = "temp.csv", net_int = NULL)
}
\arguments{
\item{net}{NHDPlus network}

\item{wbd}{WBD HU12 polygons}

\item{simp}{simplification in units of WBD and NHDPlus to limit memory usage}

\item{temp_dir}{directory to write temprary files to}

\item{out_dir}{directory to check for cached output.}
}
\description{
Parallel match levelpaths
}
\details{
Calls match_levelpaths in a parallel mode for use in large process workflows
}
